Add CID notification via Blockstore
Blockstore allows to get all keys (which are CIDs) from a chan.
Unfortunately only the first value seems to be a valid CID and
the chan keeps firing afterwards with a CID which seems not to
be valid as its string representation is only a "b". Also the
first CID has after write out to disc not the readable text
data which got added to the IPFS. Needs more investigation